set max grid size to 100
add an if function to test for an input between 1
and 100 in the prompt.

+++ index.js
@@ -27,7 +27,9 @@ function deleteGrid() {
 }
 
 gridSizeBtn.addEventListener('click', () => {
-  const gridSize = prompt('select the number of squares per side', '')
-  deleteGrid();
-  createGrid(gridSize)
+  const gridSize = prompt('select the number of squares per side (max: 100)', '')
+  if(gridSize > 0 && gridSize <= 100) {
+    deleteGrid();
+    createGrid(gridSize)
+  }
 })
